+++ src/utils/book.tsx
@@ -1,9 +1,7 @@
 import * as React from "react";
-import { BookData, FetchErrorData, BookMedium } from "interfaces";
+import { BookData, FetchErrorData, BookMedium, RequiredKeys } from "interfaces";
 import { BookFulfillmentState } from "interfaces";
-
 import { Book, Headset } from "../icons";
-import { getMedium } from "owc/utils/book";
 
 export function getAuthors(book: BookData, lim?: number): string[] {
   // select contributors if the authors array is undefined or empty.
@@ -127,3 +125,59 @@ export const bookMediumMap: {
   "http://schema.org/EBook": { name: "eBook", icon: Book },
   "http://schema.org/Book": { name: "Book", icon: Book }
 };
+
+/**
+ *  A collection of utils for processing book data
+ */
+
+export function bookIsReserved(book: BookData) {
+  return book.availability?.status === "reserved";
+}
+
+export function bookIsReady(book: BookData) {
+  return book.availability?.status === "ready";
+}
+
+export function bookIsBorrowed(
+  book: BookData
+): book is RequiredKeys<BookData, "fulfillmentLinks"> {
+  return (book.fulfillmentLinks?.length ?? 0) > 0;
+}
+
+export function bookIsOpenAccess(
+  book: BookData
+): book is RequiredKeys<BookData, "openAccessLinks"> {
+  return (book.openAccessLinks?.length ?? 0) > 0;
+}
+
+export function bookIsBorrowable(
+  book: BookData
+): book is RequiredKeys<BookData, "borrowUrl"> {
+  return typeof book.borrowUrl === "string";
+}
+
+export function getMedium(book: BookData): BookMedium | "" {
+  if (!book.raw || !book.raw["$"] || !book.raw["$"]["schema:additionalType"]) {
+    return "";
+  }
+
+  return book.raw["$"]["schema:additionalType"].value
+    ? book.raw["$"]["schema:additionalType"].value
+    : "";
+}
+
+export function getMediumSVG(
+  medium: BookMedium | "",
+  displayLabel = true
+): React.ReactNode | null {
+  if (!medium || Object.keys(bookMediumMap).indexOf(medium) === -1) {
+    return null;
+  }
+  const mediumInfo = bookMediumMap[medium];
+
+  return (
+    <div className="item-icon">
+      {mediumInfo.icon} {displayLabel ? mediumInfo.name : null}
+    </div>
+  );
+}
